feat: movie facade
