+++ spec/facades/movie_facade_spec.rb
@@ -1,63 +1,14 @@
 require 'rails_helper'
 
-RSpec.describe 'MovieFacade', :vcr do
-  before(:all) do
-    @movie_facade = MovieFacade.new
+RSpec.describe MovieFacade do
+  before :each do
+    @movie = MoviesFacade.new.get_movie(238)
+    @movie_facade = MovieFacade.new(@movie.id)
   end
 
-  describe 'get results' do
-    it "creates Movie poros for either top rated or searched" do
-      top_rated = @movie_facade.get_results('top%20rated')
-      searched = @movie_facade.get_results('searched')
-      nada = @movie_facade.get_results('')
+  it 'initializes', :vcr do
+    expect(@movie_facade).to be_a(MovieFacade)
 
-      expect(top_rated).to be_an(Array)
-      expect(top_rated).to all(be_a(Movie))
-      expect(top_rated.count).to be <= 20
-
-      expect(searched).to be_an(Array)
-      expect(searched).to all(be_a(Movie))
-      expect(searched.count).to be <= 20
-
-      expect(nada).to be_an(Array)
-      expect(nada.empty?).to eq(true)
-    end
-  end
-
-  describe 'top_rated_movies' do
-    it "creates Movie poros for top rated movies" do
-      top_movies = @movie_facade.top_rated_movies
-
-      expect(top_movies).to be_an(Array)
-      expect(top_movies).to all(be_a(Movie))
-      expect(top_movies.count).to be <= 20
-    end
-  end
-
-  describe 'search_by_title' do 
-    it "creates Movie poros for search_by_title" do
-      ocean_movies = @movie_facade.search_by_title('ocean')
-
-      expect(ocean_movies).to be_an(Array)
-      expect(ocean_movies).to all(be_a(Movie))
-      expect(ocean_movies.count).to be <= 20
-    end
-  end
-
-  describe 'get_movie' do
-    it 'creates a Movie poro for a single movie by id' do
-      movie = @movie_facade.get_movie(343611)
-
-      expect(movie).to be_a(Movie)
-    end
-  end
-
-  describe 'get_cast' do
-    it 'returns the cast of a movie that matches given id' do
-      cast = @movie_facade.get_cast('5')
-
-      expect(cast).to be_an(Array)
-      expect(cast.count).to eq(10)
-    end
+    expect(@movie_facade.movie[:title]).to eq('The Godfather')
   end
 end
